Fix miscellaneous medium/low severity bugs across multiple subsystems
- M-01: status.h - STATUS_JSON_ENCODE_ERROR duplicated STATUS_SENSOR_CREATE_ERROR (both=6); assign unique value 17
- M-02: tOutputProcess - GetOutputStateMap/GetOutputTimersStateMap returned uint8_t, truncating outputs 8-15; widened to uint16_t with (uint16_t)1<<i
- H-02: tInputProcess.h - mLastClickTime(0) caused first button press to appear as double-click; initialize to 255
- M-03: tInputProcess.cpp - mLastClickTime not reset after double-click detection, allowing spurious second doubleclick
- H-15: tSht3Sensor.cpp - logged pResult->Temperature instead of SensorID after "SensorID:" label
- M-17: lib/Random.h - operator precedence bug: (c + (b>>1)^a) evaluated as (c+(b>>1))^a; fixed to (c+((b>>1)^a))
- H-30: lib/Random.h - default constructor left a,b,c,x uninitialized; added zero initializers
- L-03: tInputProcess.cpp - bitwise | instead of || for event dispatch condition
- L-04: tOutputProcess.h - tOutput constructor left mTimer, mState, mPolarity uninitialized
- M-09: tOutputStateSensor.cpp - 1<<i on 16-bit int UB for i=15; cast to uint16_t
- M-15: NodeScanTask.cpp - 1<<(SenderDevId-1) used signed shift, no bounds check; use (uint32_t)1 with bounds

+++ tOutputProcess.cpp
@@ -189,25 +189,25 @@ void tOutputProcess::service()
 }
 
 
-uint8_t  tOutputProcess::GetOutputStateMap()
+uint16_t tOutputProcess::GetOutputStateMap()
 {
-  uint8_t Map = 0;
+  uint16_t Map = 0;
   for (uint8_t i = 0; i < getNumOfOutputs(); i++)
   {
     if (Output[i].GetState())
-      Map |= 1 << i;
+      Map |= (uint16_t)1 << i;
   }
 
   return Map;
 }
 
-uint8_t  tOutputProcess::GetOutputTimersStateMap()
+uint16_t tOutputProcess::GetOutputTimersStateMap()
 {
-  uint8_t Map = 0;
+  uint16_t Map = 0;
   for (uint8_t i = 0; i < getNumOfOutputs(); i++)
   {
     if (Output[i].GetTimer())
-      Map |= 1 << i;
+      Map |= (uint16_t)1 << i;
   }
 
   return Map;
